Add an environment file
This env was created with
$ conda create -n nlp python=3.8 anaconda
and all the specific packages were added with conda or pip

The resulting environmnent obviously contains many redundant libs
but it should at least work on a Mac/Linux machine.

Beyond that, people should be able to create an equivalent env by
following the ModuleNotFound errors.